+++ protected/humhub/modules/admin/grid/UserActionColumn.php
@@ -31,13 +31,17 @@ class UserActionColumn extends ActionColumn
             $actions[Yii::t('AdminModule.user', 'Permanently delete')] = ['delete'];
         } else {
             $actions[Yii::t('base', 'Edit')] = ['edit'];
-            $actions[] = '---';
-            if ($model->status == User::STATUS_DISABLED) {
-                $actions[Yii::t('AdminModule.user', 'Enable')] = ['enable', 'linkOptions' => ['data-method' => 'post', 'data-confirm' => Yii::t('AdminModule.user', 'Are you really sure that you want to enable this user?')]];
-            } elseif ($model->status == User::STATUS_ENABLED) {
-                $actions[Yii::t('AdminModule.user', 'Disable')] = ['disable', 'linkOptions' => ['data-method' => 'post', 'data-confirm' => Yii::t('AdminModule.user', 'Are you really sure that you want to disable this user?')]];
+
+            if(Yii::$app->user->isAdmin() || !$model->isSystemAdmin()) {
+                $actions[] = '---';
+                if ($model->status == User::STATUS_DISABLED) {
+                    $actions[Yii::t('AdminModule.user', 'Enable')] = ['enable', 'linkOptions' => ['data-method' => 'post', 'data-confirm' => Yii::t('AdminModule.user', 'Are you really sure that you want to enable this user?')]];
+                } elseif ($model->status == User::STATUS_ENABLED) {
+                    $actions[Yii::t('AdminModule.user', 'Disable')] = ['disable', 'linkOptions' => ['data-method' => 'post', 'data-confirm' => Yii::t('AdminModule.user', 'Are you really sure that you want to disable this user?')]];
+                }
+                $actions[Yii::t('base', 'Delete')] = ['delete'];
             }
-            $actions[Yii::t('base', 'Delete')] = ['delete'];
+
 
             if ($model->status == User::STATUS_ENABLED) {
                 $actions[] = '---';