+++ protected/humhub/docs/CHANGELOG.md
@@ -18,6 +18,7 @@ HumHub Change Log
 - Fix: introduction tour not working
 - Chng: Moved static js dependencies from `static/resources` into `static/js`
 - Fix: Comment edit cancel edit context item not shown
+- Fix  #2700: Prevent GroupManager access to system admin group management
 
 1.3.0-beta.2  (July 18, 2018)
 -----------------------------